requirements.txt
matplotlib==3.1.2
numpy==1.18.4
scipy==1.4.1
tensorflow_datasets==1.0.2
tensorflow-gpu==2.2.0
tensorflow-probability==0.7.0
